backend/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==4.2.9
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
